pyproject.toml
[project]
name = "recruiting_agent"
version = "1.0.0"
description = "A modular voice agent for recruitment screening calls using LiveKit, Deepgram STT, OpenAI LLM, and Cartesia TTS"
authors = [
    {name = "Hassan Aftab", email = "hassannaftabb@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "livekit-agents>=1.2.15",
    "livekit-plugins-cartesia>=1.2.15",
    "livekit-plugins-deepgram>=1.2.15",
    "livekit-plugins-silero>=1.2.15",
    "livekit-plugins-turn-detector>=1.2.15",
    "livekit-plugins-noise-cancellation>=0.2.5",
    "livekit-plugins-openai>=1.2.15",
    "pydantic>=2.12.2",
    "pydantic-settings>=2.11.0",
    "python-dotenv>=1.1.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agent"]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 88

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
